requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
PyJWT~=2.8.0
django-cors-headers==4.3.1
whitenoise==5.3.0
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.3.1
uritemplate==4.1.1
urllib3==2.2.1
webauthn==2.1.0
psycopg2-binary==2.9.9
pillow==10.4.0
